Specify websocket events structs
